+++ src/commands/create/layers-composition/layer-composition-service.ts
@@ -84,6 +84,7 @@ const defaultLayerRegistry: LayerRegistry = {
 
 const defaultRuntimeLayers: Record<string, RuntimeLayerData | undefined> = {
   "base/node": baseNodeLayer,
+  "base/static": baseStaticLayer,
 };
 
 const defaultFrameworkLayers: Record<string, FrameworkLayerData | undefined> = typedFrameworkLayers;